Second screen demo added

+++ app/src/main/java/com/example/weathertestapp/data/model/ForecastRemote.kt
@@ -27,9 +27,9 @@ data class Main(
     val humidity: Int,
     val pressure: Int,
     val sea_level: Int,
-    val temp: Int,
+    val temp: Double,
     val temp_kf: Double,
-    val temp_max: Int,
+    val temp_max: Double,
     val temp_min: Double
 )
